More changes to make plWinDPI modular
