+++ Sources/Plasma/Apps/plClient/CMakeLists.txt
@@ -76,13 +76,8 @@ set(plClient_TEXT
 )
 
 if(WIN32)
-    list(APPEND plClient_HEADERS
-        win32/plWinDpi.h
-    )
-
     list(APPEND plClient_SOURCES
         win32/plClient_Win.cpp
-        win32/plWinDpi.cpp
         win32/winmain.cpp
     )
 
@@ -151,6 +146,7 @@ target_link_libraries(
         plStatGather
         plStatusLog
         plUnifiedTime
+        $<$<BOOL:${WIN32}>:plWinDpi>
         pfAnimation
         pfAudio
         pfCharacter
