+++ Sources/Plasma/Apps/plUruLauncher/winmain.cpp
@@ -47,6 +47,8 @@ Mead, WA   99021
 #include "pfPatcher/plManifests.h"
 #include "pfPatcher/pfPatcher.h"
 
+#include "plWinDpi/plWinDpi.h"
+
 #include "plClientLauncher.h"
 
 #include "hsWindows.h"
@@ -125,6 +127,11 @@ static inline void IShowMarquee(bool marquee=true)
 
 INT_PTR CALLBACK PatcherDialogProc(HWND hwndDlg, UINT uMsg, WPARAM wParam, LPARAM lParam)
 {
+    // DPI Helper can eat messages.
+    auto result = plWinDpi::Instance().WndProc(hwndDlg, uMsg, wParam, lParam, nullptr);
+    if (result.has_value())
+        return result.value();
+
     // NT6 Taskbar Majick
     if (uMsg == s_taskbarCreated) {
         hsRequireCOM();
@@ -187,44 +194,6 @@ static void PumpMessages()
 
         // Now we need to pump the netcore while we have some spare time...
     } while (s_launcher->PumpNetCore());
-}
-
-void SetDPIAwareness() const
-{
-    plOptionalWinCall<BOOL(DPI_AWARENESS_CONTEXT)> fSetProcessDpiAwarenessContext(L"user32", "SetProcessDpiAwarenessContext");
-    plOptionalWinCall<HRESULT(PROCESS_DPI_AWARENESS)> fSetProcessDpiAwareness(L"user32", "SetProcessDpiAwarenessContext");
-
-    // There are three different levels of DPI awareness that we can deal with,
-    // each was added in newer versions of Windows as monitors became more and
-    // more sexy.
-    // In Vista, the entire system had an adjustable display scale.
-    // In Windows 8, each monitor was allowed to have a separate scale.
-    // In Windows 10, the per-monitor scaling was slowly improved to automatically
-    // scale dialog boxes and things like that.
-    // So we need to try to handle all of that. We will assume the baseline is Vista.
-    // Any API added later than that will need to be called on a provisional basis.
-    do {
-        // Per Monitor V2 for Windows 10 v1703.
-        auto perMonitorV2Result = fSetProcessDpiAwarenessContext(DPI_AWARENESS_CONTEXT_PER_MONITOR_AWARE_V2);
-        if (perMonitorV2Result.has_value()) {
-            if (perMonitorV2Result.value() == TRUE) {
-                break;
-            }
-        }
-
-        // Per Monitor v1 for Windows 8.1.
-        auto perMonitorV1Result = fSetProcessDpiAwareness(PROCESS_PER_MONITOR_DPI_AWARE);
-        if (perMonitorV1Result.has_value()) {
-            if (SUCCEEDED(perMonitorV1Result.value())) {
-                break;
-            }
-        }
-
-        // System DPI Awareness.
-        if (SetProcessDPIAware() != FALSE) {
-            break;
-        }
-    } while (false);
 }
 
 // ===================================================
@@ -419,7 +388,7 @@ static pfPatcher* IPatcherFactory()
 
 int __stdcall WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLink, int nCmdShow)
 {
-    SetProcessDPIAware();
+    plWinDpi::Instance();
 
     plClientLauncher launcher;
     s_launcher = &launcher;
