+++ Sources/Plasma/PubUtilLib/CMakeLists.txt
@@ -45,3 +45,4 @@ add_subdirectory(plSurface)
 add_subdirectory(plTransform)
 add_subdirectory(plUnifiedTime)
 add_subdirectory(plVault)
+add_subdirectory(plWinDpi)
