+++ Sources/Plasma/PubUtilLib/plMessage/plDisplayScaleChangedMsg.h
@@ -80,6 +80,18 @@ public:
 public:
     float GetScale() const { return fScale; }
     std::optional<ClientWindow> GetSuggestedLocation() const { return fNewLocation; }
+
+#if WIN32
+    static RECT ConvertRect(const plDisplayScaleChangedMsg::ClientWindow& rect)
+    {
+        return *((const LPRECT)&rect);
+    }
+
+    static plDisplayScaleChangedMsg::ClientWindow ConvertRect(const RECT& rect)
+    {
+        return *((const plDisplayScaleChangedMsg::ClientWindow*)&rect);
+    }
+#endif
 };
 
 #endif
