+++ Sources/Plasma/PubUtilLib/plWinDpi/CMakeLists.txt
@@ -1,0 +1,22 @@
+if(WIN32)
+
+list(APPEND plWinDpi_HEADERS
+    plWinDpi.h
+)
+
+list(APPEND plWinDpi_SOURCES
+    plWinDpi.cpp
+)
+
+plasma_library(plWinDpi
+    SOURCES ${plWinDpi_SOURCES} ${plWinDpi_HEADERS}
+)
+
+target_link_libraries(
+    plWinDpi
+    PUBLIC
+        CoreLib
+        plStatusLog
+)
+
+endif()
